+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/Settings/PlayerPrefsSettingsByKey.cs
@@ -15,54 +15,54 @@ namespace DCL.SettingsCommon
 
         public T GetEnum<T>(string fieldName, T defaultValue) where T : struct
         {
-            if (!Enum.TryParse<T>(PlayerPrefsUtils.GetString(GetFieldKey(fieldName), ""), out var result))
+            if (!Enum.TryParse<T>(PlayerPrefsBridge.GetString(GetFieldKey(fieldName), ""), out var result))
                 return defaultValue;
             return result;
         }
 
         public bool GetBool(string fieldName, bool defaultValue)
         {
-            return PlayerPrefsUtils.GetBool(GetFieldKey(fieldName), defaultValue);
+            return PlayerPrefsBridge.GetBool(GetFieldKey(fieldName), defaultValue);
         }
 
         public float GetFloat(string fieldName, float defaultValue)
         {
-            return PlayerPrefsUtils.GetFloat(GetFieldKey(fieldName), defaultValue);
+            return PlayerPrefsBridge.GetFloat(GetFieldKey(fieldName), defaultValue);
         }
 
         public int GetInt(string fieldName, int defaultValue)
         {
-            return PlayerPrefsUtils.GetInt(GetFieldKey(fieldName), defaultValue);
+            return PlayerPrefsBridge.GetInt(GetFieldKey(fieldName), defaultValue);
         }
 
         public string GetString(string fieldName, string defaultValue)
         {
-            return PlayerPrefsUtils.GetString(GetFieldKey(fieldName), defaultValue);
+            return PlayerPrefsBridge.GetString(GetFieldKey(fieldName), defaultValue);
         }
 
         public void SetBool(string fieldName, bool value)
         {
-            PlayerPrefsUtils.SetBool(GetFieldKey(fieldName), value);
+            PlayerPrefsBridge.SetBool(GetFieldKey(fieldName), value);
         }
 
         public void SetFloat(string fieldName, float value)
         {
-            PlayerPrefsUtils.SetFloat(GetFieldKey(fieldName), value);
+            PlayerPrefsBridge.SetFloat(GetFieldKey(fieldName), value);
         }
 
         public void SetEnum<T>(string fieldName, T value) where T : struct
         {
-            PlayerPrefsUtils.SetString(GetFieldKey(fieldName), value.ToString());
+            PlayerPrefsBridge.SetString(GetFieldKey(fieldName), value.ToString());
         }
 
         public void SetInt(string fieldName, int value)
         {
-            PlayerPrefsUtils.SetInt(GetFieldKey(fieldName), value);
+            PlayerPrefsBridge.SetInt(GetFieldKey(fieldName), value);
         }
 
         public void SetString(string fieldName, string value)
         {
-            PlayerPrefsUtils.SetString(GetFieldKey(fieldName), value);
+            PlayerPrefsBridge.SetString(GetFieldKey(fieldName), value);
         }
 
         private string GetFieldKey(string fieldName)
