+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Helpers/Utils/PlayerPrefsUtils/IPlayerPrefsProvider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eab9626a4b5dd48e392a598c9033c830
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
