+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Helpers/Utils/PlayerPrefsUtils/PlayerPrefsProviderDefault.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cba97ba6d1a6c424a8c348e9082ae53e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
